Correcoes

+++ cuda/host.cu
@@ -62,7 +62,7 @@ double box_filter_8u_c1(initialParams* ct, PPMImageParams* imageParams, PPMBlock
 {
 
     cudaEvent_t start, stop;
-    double time;
+    float time;
     cudaEventCreate(&start);
     cudaEventCreate(&stop);
 
@@ -139,7 +139,7 @@ double box_filter_8u_c1(initialParams* ct, PPMImageParams* imageParams, PPMBlock
     cudaFree(GPU_output);
 
     cudaEventElapsedTime(&time, start, stop);
-    return time;
+    return (double)time;
 }
 
 // FUNCAO __HOST__
